Split debug selector tests
